+++ DEV-1/DEV-1/DEV-1.csproj
@@ -40,8 +40,8 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="IdenticalSymbols.cs" />
-    <Compile Include="Program.cs" />
+    <Compile Include="SerialIdenticalSymbols.cs" />
+    <Compile Include="EntryPoint.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
